# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Leverpay.io API Documentation",
        "description": "Crypto payment gateway",
        "contact": {
            "email": "abdilkura@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.leverpay.io",
            "description": "Test server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/add-payment-option": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Create new payment option",
                "operationId": "create payment option",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "icon"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "icon": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/admin/get-all-merchants": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all merchants",
                "operationId": "get all merchants",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/admin/get-all-users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all user",
                "operationId": "get all user",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/admin/get-kyc-list": {
            "get": {
                "tags": [
                    "KYC"
                ],
                "summary": "Get all KYC details",
                "operationId": "Get all KYC",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/v1/kyc-details": {
            "get": {
                "tags": [
                    "KYC"
                ],
                "summary": "Get all kyc details",
                "operationId": "Get all kyc details",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://api.leverpay.io/oauth/authorize",
                        "tokenUrl": "https://api.leverpay.io/oauth/token",
                        "refreshUrl": "https://api.leverpay.io/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    ]
}